Doi ten controller
